chrona-backend/pyproject.toml
[project]
name = "chrona-backend"
version = "2.0.0"
description = "A sleek, efficient, customizable countdown timer with smooth animations."
authors = [
    {name = "Miracle Mayowa Adebunmi",email = "blueprime91@gmail.com"}
]
license = {text = "Proprietary"}
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "django (==5.1.7)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "django-storages (>=1.14.6,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "boto3 (>=1.37.29,<2.0.0)",
    "social-auth-app-django (>=5.4.3,<6.0.0)",
    "social-auth-core (>=4.5.6,<5.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "python-json-logger (>=3.3.0,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "django-prometheus (>=2.3.1,<3.0.0)",
    "argon2-cffi (>=23.1.0,<24.0.0)",
    "inflection (>=0.5.1,<0.6.0)",
    "celery (>=5.5.1,<6.0.0)",
    "django-celery-beat (>=2.8.0,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "azure-identity (>=1.21.0,<2.0.0)",
    "azure-keyvault-secrets (>=4.9.0,<5.0.0)",
    "azure-mgmt-keyvault (>=11.0.0,<12.0.0)",
    "azure-mgmt-resource (>=23.3.0,<24.0.0)",
    "pywebpush (>=2.0.3,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-django = "^4.11.1"
pytest-cov = "^6.1.1"
factory-boy = "^3.3.3"

